[feat] math_func: derivative

+++ geometry/include/geom_utils.h
@@ -9,7 +9,7 @@ static inline Point2i toImageCoord(const Point2f p, const Point2i origin) {
 }
 
 static inline Point2f toMathCoord(const Point2i p, const Point2i origin) {
-    return (Point2f){p.x - origin.x, origin.y - p.y};
+    return (Point2f){(float)(p.x - origin.x), (float)(origin.y - p.y)};
 }
 
 static inline float minf(const float a, const float b) {
